requirements.txt
discord.py==2.8.0